fixed modal window button behavior when deleting a category

+++ src/containers/edit-profile/professional-info-tab/professional-category/ProfessionalCategory.tsx
@@ -45,8 +45,8 @@ const ProfessionalCategory: FC<ProfessionalCategoryProps> = ({
       confirmButton: t(
         'editProfilePage.profile.professionalTab.deleteCategoryModal.submitBtn'
       ),
-      sendConfirm: () => {
-        handleDelete()
+      sendConfirm: (isConfirmed) => {
+        if (isConfirmed) handleDelete()
       }
     })
   }
